Refactor convention checkers
Does the following for the checkers -
 * Improves re compile. Currently we were doing a re compile inside
function calls. Moves this out so that this compile only happens once.
 * Improve checker code quality -
   * Current checker is stateless. It was still written as a class which
	added additional complexity with addition of passing of `self`
	`cls` and using staticmethod decorators.
   * Move the checker hook out of the class.
   * Slice up the checks into base, other and numpy style checks.
